Fixes issue #2762 - Move all CoreProfile integration tests into /test/coreprofile sub directory
