+++ test/coreprofile/helloarquillian/src/main/webapp/WEB-INF/beans.xml
