+++ test/coreprofile/helloworld/src/main/webapp/WEB-INF/beans.xml
